Added: Program For Calculating Area Of Circle

+++ basic-programs/AreaOfCircle.java
@@ -3,7 +3,7 @@ public class AreaOfCircle{
 	public static void main(String[] args){
 		System.out.println("Enter radius: ");
 		Scanner sc = new Scanner(System.in);
-		int r = sc.nextInt();
+		double r = sc.nextDouble();
 		double Area = 3.14 * r * r;
 		System.out.println("Area of circle is: "+ Area);
 		sc.close();
